moves post and reply features specs into their own files
